migrating changes from #19184
