+++ ui/components/app/signature-request/index.scss
@@ -22,7 +22,8 @@
   }
 
   &__reject-all-button {
-    margin-top: -15px;
+    margin-bottom: -5px;
+    padding-bottom: 20px;
   }
 
   &__origin {
