+++ ui/components/app/signature-request/signature-request.test.js
@@ -4,7 +4,7 @@ import { fireEvent } from '@testing-library/react';
 import configureMockStore from 'redux-mock-store';
 import mockState from '../../../../test/data/mock-state.json';
 import { renderWithProvider } from '../../../../test/lib/render-helpers';
-import { SECURITY_PROVIDER_MESSAGE_SEVERITIES } from '../security-provider-banner-message/security-provider-banner-message.constants';
+import { SECURITY_PROVIDER_MESSAGE_SEVERITY } from '../../../../shared/constants/security-provider';
 import {
   getNativeCurrency,
   getProviderConfig,
@@ -388,8 +388,7 @@ describe('Signature Request Component', () => {
           txData={{
             msgParams,
             securityProviderResponse: {
-              flagAsDangerous:
-                SECURITY_PROVIDER_MESSAGE_SEVERITIES.NOT_MALICIOUS,
+              flagAsDangerous: SECURITY_PROVIDER_MESSAGE_SEVERITY.NOT_MALICIOUS,
             },
           }}
         />,
@@ -437,8 +436,7 @@ describe('Signature Request Component', () => {
           txData={{
             msgParams,
             securityProviderResponse: {
-              flagAsDangerous:
-                SECURITY_PROVIDER_MESSAGE_SEVERITIES.NOT_MALICIOUS,
+              flagAsDangerous: SECURITY_PROVIDER_MESSAGE_SEVERITY.NOT_MALICIOUS,
             },
           }}
         />,
